+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @ai-sdk/angular
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [8e72304]
+  - ai@5.0.7
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [d983eee]
+  - ai@5.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- ai@5.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- ai@5.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [90d212f]
+  - @ai-sdk/provider-utils@3.0.1
+  - ai@5.0.3
+
 ## 1.0.2
 
 ### Patch Changes