+++ packages/google/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/google
 
+## 2.0.3
+
+### Patch Changes
+
+- 9fb0252: fix(google): add thought signature support for reasoning
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [90d212f]
+  - @ai-sdk/provider-utils@3.0.1
+
 ## 2.0.1
 
 ### Patch Changes