+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,39 @@
 
 ### Patch Changes
 
+- Updated dependencies [8e72304]
+  - ai@5.0.7
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [d983eee]
+  - ai@5.0.6
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.5
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.4
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [90d212f]
+  - ai@5.0.3
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [401d73e]
 - Updated dependencies [69fde99]
   - ai@5.0.2